api/requirements.txt
pandas==2.0.2
geopandas==0.13.2
psycopg2==2.9.6
sqlalchemy==2.0.15
python-dotenv==1.0.0